feat: move the TileSpecifications to a config folder
