+++ src/track/config/index.js
@@ -16,8 +16,4 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-export { CurvedTileEnlargedModel } from './CurvedTileEnlargedModel.js';
-export { CurvedTileModel } from './CurvedTileModel.js';
-export { StraightTileModel } from './StraightTileModel.js';
-export { TileModel } from './TileModel.js';
-export { TrackModel } from './TrackModel.js';
+export { TileSpecifications } from './TileSpecifications.js';
