+++ src/track/models/test/CurvedTileEnlargedModel.spec.js
@@ -17,7 +17,7 @@
  */
 
 import { CurvedTileEnlargedModel } from '../CurvedTileEnlargedModel.js';
-import { TileSpecifications } from '../TileSpecifications.js';
+import { TileSpecifications } from '../../config';
 
 const tileX = 100;
 const tileY = 100;
