+++ src/track/models/test/CurvedTileModel.spec.js
@@ -17,7 +17,7 @@
  */
 
 import { CurvedTileModel } from '../CurvedTileModel.js';
-import { TileSpecifications } from '../TileSpecifications.js';
+import { TileSpecifications } from '../../config';
 
 const tileX = 100;
 const tileY = 100;
