+++ src/track/models/test/StraightTileModel.spec.js
@@ -17,7 +17,7 @@
  */
 
 import { StraightTileModel } from '../StraightTileModel.js';
-import { TileSpecifications } from '../TileSpecifications.js';
+import { TileSpecifications } from '../../config';
 
 const tileX = 100;
 const tileY = 100;
